made middlewares more predictable
as router is now not a two-layer tree, but n-layer, middlewares are now easier to implement correctly (and with determined call chain). Btw about call chains - made them static & eliminated recursion, so they are even faster now
